feat: Add ChatRoomList component
This commit adds a new component `ChatRoomList.tsx` to the `src/pages` directory. The `ChatRoomList` component displays a list of chat rooms fetched from the server. It includes an `AppBar` component, a title, and a list of `RoomTile` components. This component will be used as the main page for the chat feature in the application.